+++ app/games/typing-speed/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+import { useApi } from "@/hooks/useApi";
+import useLocalStorage from "@/hooks/useLocalStorage";
+import { User } from "@/types/user";
+import { useParams, useRouter } from "next/navigation";
+
+
+import React, { useEffect, useState } from "react";
+
